+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Enhancements:
 
 - feat(client): Add support for impersonation. [#724](https://github.com/fastly/go-fastly/pull/724)
-- feat(ngwaf/v1/rules): CRUD for rate limit workspace rules [#727](https://github.com/fastly/go-fastly/pull/727)
+- feat(ngwaf/v1/rules): Add support for rate limit workspace rules [#727](https://github.com/fastly/go-fastly/pull/727)
 
 ### Bug fixes: