+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [UNRELEASED]
 
-[Full Changelog](https://github.com/fastly/go-fastly/compare/v11.1.0...)
+[Full Changelog](https://github.com/fastly/go-fastly/compare/v11.1.1...)
 
 ### Breaking:
 
@@ -10,9 +10,15 @@
 
 ### Bug fixes:
 
-- fix(ngwaf/v1/rules): Fix Typing on rate limit struct. [#729](https://github.com/fastly/go-fastly/pull/729)
-
 ### Dependencies:
+
+## [v11.1.1](https://github.com/fastly/go-fastly/releases/tag/v11.1.1) (2025-08-04)
+
+[Full Changelog](https://github.com/fastly/go-fastly/compare/v11.1.0...v11.1.1)
+
+### Bug fixes:
+
+- fix(ngwaf/v1/rules): Fix Typing on rate limit struct. [#729](https://github.com/fastly/go-fastly/pull/729)
 
 ## [v11.1.0](https://github.com/fastly/go-fastly/releases/tag/v11.1.0) (2025-08-04)